+++ src/main/java/com/souma1024/shogiv2/config/WebSocketConfig.java
@@ -7,6 +7,7 @@ import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
 import com.souma1024.shogiv2.repository.RoomRepository;
 import com.souma1024.shogiv2.service.GameStartService;
+import com.souma1024.shogiv2.service.MovablePositionService;
 import com.souma1024.shogiv2.service.RoomSessionManager;
 import com.souma1024.shogiv2.websocket.ShogiWebSocketHandler;
 
@@ -17,15 +18,17 @@ public class WebSocketConfig implements WebSocketConfigurer {
     private final GameStartService gameStartService;
     private final RoomRepository roomRepository;
     private final RoomSessionManager roomManager;
+    private final MovablePositionService movablePositionService;
 
-    WebSocketConfig(RoomRepository roomRepository, GameStartService gameStartService, RoomSessionManager roomManager) {
+    WebSocketConfig(RoomRepository roomRepository, GameStartService gameStartService, RoomSessionManager roomManager, MovablePositionService movablePositionService) {
         this.roomRepository = roomRepository;
         this.gameStartService = gameStartService; 
         this.roomManager = roomManager;
+        this.movablePositionService = movablePositionService;
     }
     @Override
     public void registerWebSocketHandlers(@NonNull WebSocketHandlerRegistry registry) {
-        registry.addHandler(new ShogiWebSocketHandler(roomRepository, gameStartService, roomManager), "/ws/shogi")
+        registry.addHandler(new ShogiWebSocketHandler(roomRepository, gameStartService, roomManager, movablePositionService), "/ws/shogi")
                 .setAllowedOrigins("*");
     }
 }
